SignUp, SignIn and SetPin Pages Added
